+++ src/const.ts
@@ -1,7 +1,7 @@
 export const CREATED = 'CREATED';
-export const MIGRATED = 'MIGRATED';
-export const MIGRATION_FAILED = 'MIGRATION_FAILED';
+export const APPLIED = 'APPLIED';
+export const APPLY_FAILED = 'APPLY_FAILED';
 export const REVERTED = 'REVERTED';
 export const REVERT_FAILED = 'REVERT_FAILED';
 
-export const PRINT_PRIORITY = ['dateReverted', 'dateMigrated', 'dateCreated'];
+export const PRINT_PRIORITY = ['dateReverted', 'dateApplied', 'dateCreated'];