+++ src/templates/migration.ts
@@ -1,11 +1,11 @@
-// import models here...
+// import or write Mongoose schemas and models here...
 
 export async function up() {
-  // write upgrade script here... throw Error or return false to fail migration.
+  // write upgrade script here... throw Error or return false to fail upgrade.
   return true;
 }
 
 export async function down() {
-  // write revert script here... throw Error or return false to fail migration.
+  // write revert script here... throw Error or return false to fail revert.
   return true;
 }