+++ src/types.ts
@@ -1,19 +1,19 @@
 import {
   CREATED,
-  MIGRATED,
-  MIGRATION_FAILED,
+  APPLIED,
+  APPLY_FAILED,
   REVERTED,
   REVERT_FAILED,
 } from './const';
 
 export type IMigration = {
   dateCreated?: Date;
-  dateMigrated?: Date;
+  dateApplied?: Date;
   dateReverted?: Date;
   name: string;
   status: typeof CREATED
-        | typeof MIGRATED
-        | typeof MIGRATION_FAILED
+        | typeof APPLIED
+        | typeof APPLY_FAILED
         | typeof REVERTED
         | typeof REVERT_FAILED
         ;