+++ CHANGELOG.md
@@ -14,6 +14,11 @@ _Released 2024/10/TBD_
 
 - A `CHANGELOG.md` file to track project changes.
 
+**Changed:**
+
+- Renamed status `MIGRATED` to `APPLIED` and `MIGRATION_FAILED` to `APPLY_FAILED`.
+- Renamed the `Migration` Model property `dateMigrated` to `dateApplied`.
+
 
 ## 0.0.1-alpha.2
 _Released 2024/10/01_